enable read_csv() for both schema dictionary and file
