+++ exetera/core/__init__.py
@@ -1,3 +1,3 @@
 
-from . import data_writer, csvdataset, exporter, fields, filtered_field, schema_importer,\
-    operations, persistence, readerwriter, regression, session, split, utils, validation
+from . import data_writer, csvdataset, exporter, fields, filtered_field, operations, \
+              persistence, readerwriter, regression, session, split, utils, validation
