+++ tests/test_json_importer.py
@@ -4,8 +4,8 @@ import tempfile
 import os
 import h5py
 from datetime import datetime, timezone
-from exetera.core import schema_importer as importer, session
-from exetera.core import utils
+from exetera.io import json_importer as importer
+from exetera.core import utils, session
 import numpy as np
 from io import BytesIO, StringIO
